Disable long double functionality where not available.
Currently long double is disabled explicitly on Cygwin where it is known not to be available (more
specifically long double functions are just aliases for the normal double versions).

This does keep the mpz_ld member of the fplll_type_t enum, even without HAVE_LONG_DOUBLE, which was
convenient in a few places, but this could be removed with a little restructuring.